# docs/swagger.yaml
basePath: /api
definitions:
  handler.AdjustBalanceRequest:
    properties:
      delta:
        example: 100
        type: integer
      remark:
        example: 管理员充值
        type: string
    required:
    - delta
    type: object
  handler.ConsumeRequest:
    properties:
      amount:
        example: 1
        type: integer
      key:
        example: CK-xxxx-xxxx
        type: string
    required:
    - key
    type: object
  handler.CreateKeyJSON:
    properties:
      alias:
        example: 测试卡密
        type: string
      remaining_amount:
        example: 100
        type: integer
    required:
    - alias
    - remaining_amount
    type: object
...
  handler.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  handler.UpdateKeyConfigRequest:
    properties:
      key_length:
        type: integer
      key_prefix:
        type: string
      key_suffix_length:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 卡密管理系统 API 文档
  title: CloudKey API
  version: "1.0"
paths:
...
  /service/keys:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: '状态过滤: active/exhausted/disabled/expired'
        in: query
        name: status
        type: string
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页卡密列表
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密列表
      tags:
      - 服务账号API
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，服务账号创建卡密
      parameters:
      - description: 卡密参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号创建卡密
      tags:
      - 服务账号API
  /service/keys/{id}:
    delete:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号删除卡密
      tags:
      - 服务账号API
    get:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 卡密详情
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密详情
      tags:
      - 服务账号API
    patch:
      consumes:
      - application/json
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号更新卡密
      tags:
      - 服务账号API
...
  /tenant/key-config:
    get:
      description: 租户管理员获取当前租户的卡密生成配置
      produces:
      - application/json
      responses:
        "200":
          description: Key 配置信息
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 租户不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取 Key 配置
      tags:
      - 租户-配置
    patch:
      consumes:
      - application/json
      description: 租户管理员更新当前租户的卡密生成配置
      parameters:
      - description: 配置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateKeyConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新 Key 配置
      tags:
      - 租户-配置
  /tenant/keys:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: '状态过滤: active/exhausted/disabled/expired'
        in: query
        name: status
        type: string
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页卡密列表
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 卡密列表
      tags:
      - 租户-卡密管理
    post:
      consumes:
      - application/json
      description: 租户管理员创建新卡密，需业务状态正常
      parameters:
      - description: 卡密参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateKeyJSON'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，含 raw_key
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建卡密
      tags:
      - 租户-卡密管理
  /tenant/keys/{id}:
    delete:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除卡密
      tags:
      - 租户-卡密管理
    get:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 卡密详情
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 卡密详情
      tags:
      - 租户-卡密管理
    patch:
      consumes:
      - application/json
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新卡密
      tags:
      - 租户-卡密管理
  /tenant/keys/{id}/adjust-balance:
    post:
      consumes:
      - application/json
      description: 管理员增加或减少卡密额度，仅支持增量操作，所有变动记录在流转日志中
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 调整参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AdjustBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 调整结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 调整卡密额度
      tags:
      - 租户-卡密管理
  /tenant/keys/{id}/disable:
    patch:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 禁用成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 禁用卡密
      tags:
      - 租户-卡密管理
  /tenant/keys/{id}/enable:
    patch:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 启用成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 启用卡密
      tags:
      - 租户-卡密管理
  /tenant/keys/consume:
    post:
      consumes:
      - application/json
      description: 扣减指定卡密的剩余额度
      parameters:
      - description: 扣减参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ConsumeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 扣减结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误或卡密无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 扣减卡密额度
      tags:
      - 租户-卡密管理
  /tenant/keys/consume-result:
    get:
      description: 租户管理员根据 request_id 查询消费或调额的结果
      parameters:
      - description: 请求ID
        in: query
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 缺少 request_id 参数
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 未找到对应的操作记录
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 根据请求ID查询操作结果
      tags:
      - 租户-卡密管理
  /tenant/keys/export:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 导出卡密（文本格式）
      tags:
      - 租户-卡密管理
  /tenant/keys/export/json:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据 JSON 数组
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 导出卡密（JSON 格式）
      tags:
      - 租户-卡密管理
  /tenant/keys/status:
    get:
      description: 根据卡密值查询卡密状态，不扣减额度
      parameters:
      - description: 卡密值
        in: query
        name: sk
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 卡密状态信息
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 缺少卡密参数
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询卡密状态
      tags:
      - 租户-卡密管理
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer JWT token, 格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  ServiceKeyAuth:
    description: 服务账号密钥
    in: header
    name: X-Service-Key
    type: apiKey
swagger: "2.0"